check for the existence of the link

+++ install.sh
@@ -23,7 +23,9 @@ function installDotFiles (){
   mkdir -p $HOME/.bash/
 
   cp files/terminator.config $HOME/.config/terminator/
-  ln -s $HOME/.config/terminator/terminator.config $HOME/.config/terminator/config
+  if ! [ -s  $HOME/.config/terminator/config ]; then
+    ln -s $HOME/.config/terminator/terminator.config $HOME/.config/terminator/config
+  fi
   cp files/git_prompt.sh $HOME/.bash/
   cp files/shell_prompt.sh $HOME/.bash/
   cp files/bashrc $HOME/.bashrc
